📝 doc: modify example code structure to add broker client examples
