--- example/components/SectionGetAsset.tsx
@@ -1,59 +1,0 @@
-import { AssetInfo } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
-import { useState } from 'react';
-import { useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import { object, string } from 'yup';
-import FormField from './FormField';
-
-const schema = object({
-  assetId: string().required()
-});
-
-export default function SectionGetAsset({ client }: { client: HibitClient }) {
-  const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<AssetInfo | null>(null);
-  const [error, setError] = useState<string>('');
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors }
-  } = useForm({
-    resolver: yupResolver(schema)
-  });
-
-  const submit = handleSubmit(async (input) => {
-    setLoading(true);
-    setResult(null);
-    setError('');
-    try {
-      const req = BigInt(input.assetId);
-      setResult(await client.getAsset(req));
-    } catch (e: any) {
-      setError(e.message ?? JSON.stringify(e));
-    } finally {
-      setLoading(false);
-    }
-  });
-
-  return (
-    <Section
-      title="GetAsset"
-      form={
-        <div className="flex flex-col gap-2">
-          <FormField label="AssetId" error={errors.assetId} required>
-            <input type="number" className="input" {...register('assetId')} />
-          </FormField>
-          <button className="btn" onClick={submit} disabled={loading}>
-            {loading ? 'Loading...' : 'Submit'}
-          </button>
-        </div>
-      }
-      loading={loading}
-      result={result}
-      error={error}
-    />
-  );
-}
