+++ example/components/hibit-client-sections/SectionGetMarketDepth.tsx
@@ -1,33 +1,27 @@
-import { GetMarketKlineInput, MarketKlineItem, PageResponse, TickSpace } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
+import { DepthIndex, GetMarketDepthInput, MarketDepth } from '../../../src';
+import { HibitClient } from '../../../src/hibit-client';
+import Section from '../Section';
 import { useState } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { object, number, string } from 'yup';
-import FormField from './FormField';
+import FormField from '../FormField';
 
 const schema = object({
-  marketId: string().required(),
-  tickSpace: number()
+  index: number()
     .oneOf(
-      Object.keys(TickSpace)
+      Object.keys(DepthIndex)
         .filter((v) => !isNaN(Number(v)))
         .map(Number)
     )
     .required(),
-  limit: number()
-    .nullable()
-    .transform((value, original) => (original === '' ? null : value)),
-  offset: number()
-    .nullable()
-    .transform((value, original) => (original === '' ? null : value)),
-  orderBy: string()
+  marketId: string().required(),
+  limit: number().required().min(1).max(100)
 });
 
-export default function SectionGetMarketKline({ client }: { client: HibitClient }) {
+export default function SectionGetMarketDepth({ client }: { client: HibitClient }) {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<PageResponse<MarketKlineItem> | null>(null);
+  const [result, setResult] = useState<MarketDepth | null>(null);
   const [error, setError] = useState<string>('');
 
   const {
@@ -36,6 +30,9 @@ export default function SectionGetMarketKline({ client }: { client: HibitClient 
     handleSubmit,
     formState: { errors }
   } = useForm({
+    defaultValues: {
+      limit: 10
+    },
     resolver: yupResolver(schema)
   });
 
@@ -44,14 +41,12 @@ export default function SectionGetMarketKline({ client }: { client: HibitClient 
     setResult(null);
     setError('');
     try {
-      const req: GetMarketKlineInput = {
+      const req: GetMarketDepthInput = {
+        index: input.index,
         marketId: BigInt(input.marketId),
-        tickSpace: input.tickSpace,
-        limit: input.limit ?? undefined,
-        offset: input.offset ?? undefined,
-        orderBy: input.orderBy || undefined
+        limit: input.limit
       };
-      setResult(await client.getMarketKline(req));
+      setResult(await client.getMarketDepth(req));
     } catch (e: any) {
       setError(e.message ?? JSON.stringify(e));
     } finally {
@@ -61,30 +56,27 @@ export default function SectionGetMarketKline({ client }: { client: HibitClient 
 
   return (
     <Section
-      title="GetMarketKline"
+      title="GetMarketDepth"
       form={
         <div className="flex flex-col gap-2">
-          <FormField label="MarketId" error={errors.marketId} required>
-            <input type="number" className="input" {...register('marketId')} />
-          </FormField>
-          <FormField label="TickSpace" error={errors.tickSpace} required>
+          <FormField label="Index" error={errors.index} required>
             <Controller
-              name="tickSpace"
+              name="index"
               control={control}
               render={({ field }) => (
                 <div className="flex items-center gap-3 flex-wrap">
-                  {Object.keys(TickSpace)
+                  {Object.keys(DepthIndex)
                     .filter((v) => isNaN(Number(v)))
                     .map((key) => (
                       <label key={key} className="flex items-center gap-1">
                         <span>{key}</span>
                         <input
-                          name="getMarketKline-tickSpace"
+                          name="getMarketDepth-index"
                           type="radio"
-                          checked={field.value === Number(TickSpace[key as any])}
+                          checked={field.value === Number(DepthIndex[key as any])}
                           onChange={(ev) => {
                             if (ev.target.checked) {
-                              field.onChange(Number(TickSpace[key as any]));
+                              field.onChange(Number(DepthIndex[key as any]));
                             }
                           }}
                         />
@@ -94,14 +86,11 @@ export default function SectionGetMarketKline({ client }: { client: HibitClient 
               )}
             />
           </FormField>
-          <FormField label="Limit" error={errors.limit}>
-            <input type="number" className="input" {...register('limit')} />
+          <FormField label="MarketId" error={errors.marketId} required>
+            <input type="number" className="input" {...register('marketId')} />
           </FormField>
-          <FormField label="Offset" error={errors.offset}>
-            <input type="number" className="input" {...register('offset')} />
-          </FormField>
-          <FormField label="OrderBy" error={errors.orderBy}>
-            <input type="text" className="input" {...register('orderBy')} />
+          <FormField label="Limit" error={errors.limit} required>
+            <input type="number" min={1} max={100} className="input" {...register('limit')} />
           </FormField>
           <button className="btn" onClick={submit} disabled={loading}>
             {loading ? 'Loading...' : 'Submit'}
