+++ example/components/hibit-client-sections/SectionGetMarkets.tsx
@@ -1,17 +1,19 @@
-import { AssetInfo, ChainId, GetAssetsInput, PageResponse } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
+import { GetMarketsInput, MarketInfo, PageResponse, ChainId } from '../../../src';
+import { HibitClient } from '../../../src/hibit-client';
+import Section from '../Section';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { object, number, string } from 'yup';
-import FormField from './FormField';
-import ChainIdSelector from './ChainIdSelector';
-import AssetTypeSelector from './AssetTypeSelector';
+import FormField from '../FormField';
+import ChainIdSelector from '../ChainIdSelector';
+import AssetTypeSelector from '../AssetTypeSelector';
 
 const schema = object({
   chainIds: object().nullable(),
   chainAssetTypes: object().nullable(),
+  baseAssetId: string(),
+  quoteAssetId: string(),
   limit: number()
     .nullable()
     .transform((value, original) => (original === '' ? null : value)),
@@ -21,9 +23,9 @@ const schema = object({
   orderBy: string()
 });
 
-export default function SectionGetAssets({ client }: { client: HibitClient }) {
+export default function SectionGetMarkets({ client }: { client: HibitClient }) {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<PageResponse<AssetInfo> | null>(null);
+  const [result, setResult] = useState<PageResponse<MarketInfo> | null>(null);
   const [error, setError] = useState<string>('');
   const [selectedAssetTypes, setSelectedAssetTypes] = useState<number[]>([]);
   const [selectedChainIds, setSelectedChainIds] = useState<ChainId[]>([]);
@@ -41,14 +43,16 @@ export default function SectionGetAssets({ client }: { client: HibitClient }) {
     setResult(null);
     setError('');
     try {
-      const req: GetAssetsInput = {
+      const req: GetMarketsInput = {
         chainIds: selectedChainIds.length > 0 ? selectedChainIds : undefined,
         chainAssetTypes: selectedAssetTypes.length > 0 ? selectedAssetTypes : undefined,
+        baseAssetId: input.baseAssetId ? BigInt(input.baseAssetId) : undefined,
+        quoteAssetId: input.quoteAssetId ? BigInt(input.quoteAssetId) : undefined,
         limit: input.limit ?? undefined,
         offset: input.offset ?? undefined,
         orderBy: input.orderBy || undefined
       };
-      setResult(await client.getAssets(req));
+      setResult(await client.getMarkets(req));
     } catch (e: any) {
       setError(e.message ?? JSON.stringify(e));
     } finally {
@@ -58,7 +62,7 @@ export default function SectionGetAssets({ client }: { client: HibitClient }) {
 
   return (
     <Section
-      title="GetAssets"
+      title="GetMarkets"
       form={
         <div className="flex flex-col gap-2">
           <ChainIdSelector selectedChainIds={selectedChainIds} onChange={setSelectedChainIds} error={errors.chainIds} />
@@ -67,6 +71,12 @@ export default function SectionGetAssets({ client }: { client: HibitClient }) {
             onChange={setSelectedAssetTypes}
             error={errors.chainAssetTypes}
           />
+          <FormField label="BaseAssetId" error={errors.baseAssetId}>
+            <input type="number" className="input" {...register('baseAssetId')} />
+          </FormField>
+          <FormField label="QuoteAssetId" error={errors.quoteAssetId}>
+            <input type="number" className="input" {...register('quoteAssetId')} />
+          </FormField>
           <FormField label="Limit" error={errors.limit}>
             <input type="number" className="input" {...register('limit')} />
           </FormField>
