+++ example/components/hibit-client-sections/SectionGetMarkets24HrTickerExtended.tsx
@@ -1,13 +1,13 @@
-import { ChainId, GetMarket24HrTickerInput, Market24HrTickerInfo } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
+import { ChainId, GetMarket24HrTickerInput, Market24HrTickerExtendInfo } from '../../../src';
+import { HibitClient } from '../../../src/hibit-client';
+import Section from '../Section';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { object, string } from 'yup';
-import FormField from './FormField';
-import ChainIdSelector from './ChainIdSelector.tsx';
-import AssetTypeSelector from './AssetTypeSelector.tsx';
+import FormField from '../FormField';
+import ChainIdSelector from '../ChainIdSelector';
+import AssetTypeSelector from '../AssetTypeSelector';
 
 const schema = object({
   marketId: string(),
@@ -15,9 +15,9 @@ const schema = object({
   chainAssetTypes: object().nullable()
 });
 
-export default function SectionGetMarkets24HrTicker({ client }: { client: HibitClient }) {
+export default function SectionGetMarkets24HrTickerExtended({ client }: { client: HibitClient }) {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<Array<Market24HrTickerInfo> | null>(null);
+  const [result, setResult] = useState<Array<Market24HrTickerExtendInfo> | null>(null);
   const [error, setError] = useState<string>('');
   const [selectedAssetTypes, setSelectedAssetTypes] = useState<number[]>([]);
   const [selectedChainIds, setSelectedChainIds] = useState<ChainId[]>([]);
@@ -40,7 +40,7 @@ export default function SectionGetMarkets24HrTicker({ client }: { client: HibitC
         chainIds: selectedChainIds.length > 0 ? selectedChainIds : undefined,
         chainAssetTypes: selectedAssetTypes.length > 0 ? selectedAssetTypes : undefined
       };
-      setResult(await client.getMarkets24HrTicker(req));
+      setResult(await client.getMarkets24HrTickerExtended(req));
     } catch (e: any) {
       setError(e.message ?? JSON.stringify(e));
     } finally {
@@ -50,7 +50,7 @@ export default function SectionGetMarkets24HrTicker({ client }: { client: HibitC
 
   return (
     <Section
-      title="GetMarkets24HrTicker"
+      title="GetMarkets24HrTickerExtended"
       form={
         <div className="flex flex-col gap-2">
           <FormField label="MarketId" error={errors.marketId}>
