+++ example/components/hibit-client-sections/SectionGetMarketsSwapInfo.tsx
@@ -1,19 +1,19 @@
-import { MarketInfo } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
+import { MarketSwapInfo } from '../../../src';
+import { HibitClient } from '../../../src/hibit-client';
+import Section from '../Section';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { object, string } from 'yup';
-import FormField from './FormField';
+import FormField from '../FormField';
 
 const schema = object({
-  marketId: string().required()
+  marketId: string()
 });
 
-export default function SectionGetMarket({ client }: { client: HibitClient }) {
+export default function SectionGetMarketsSwapInfo({ client }: { client: HibitClient }) {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<MarketInfo | null>(null);
+  const [result, setResult] = useState<Array<MarketSwapInfo> | null>(null);
   const [error, setError] = useState<string>('');
 
   const {
@@ -29,8 +29,8 @@ export default function SectionGetMarket({ client }: { client: HibitClient }) {
     setResult(null);
     setError('');
     try {
-      const req = BigInt(input.marketId);
-      setResult(await client.getMarket(req));
+      const req = input.marketId ? BigInt(input.marketId) : undefined;
+      setResult(await client.getMarketsSwapInfo(req));
     } catch (e: any) {
       setError(e.message ?? JSON.stringify(e));
     } finally {
@@ -40,10 +40,10 @@ export default function SectionGetMarket({ client }: { client: HibitClient }) {
 
   return (
     <Section
-      title="GetMarket"
+      title="GetMarketsSwapInfo"
       form={
         <div className="flex flex-col gap-2">
-          <FormField label="MarketId" error={errors.marketId} required>
+          <FormField label="MarketId" error={errors.marketId}>
             <input type="number" className="input" {...register('marketId')} />
           </FormField>
           <button className="btn" onClick={submit} disabled={loading}>
