+++ example/components/hibit-client-sections/SectionGetNonce.tsx
@@ -1,21 +1,18 @@
-import { GetWalletBalancesInput } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
+import { HibitClient } from '../../../src/hibit-client';
+import Section from '../Section';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { object, string } from 'yup';
-import FormField from './FormField';
-import BigNumber from 'bignumber.js';
+import FormField from '../FormField';
 
 const schema = object({
-  hin: string().required(),
-  assetId: string()
+  hin: string().required()
 });
 
-export default function SectionGetWalletBalances({ client }: { client: HibitClient }) {
+export default function SectionGetNonce({ client }: { client: HibitClient }) {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<Map<string, BigNumber> | null>(null);
+  const [result, setResult] = useState<bigint | null>(null);
   const [error, setError] = useState<string>('');
 
   const {
@@ -31,12 +28,8 @@ export default function SectionGetWalletBalances({ client }: { client: HibitClie
     setResult(null);
     setError('');
     try {
-      const req: GetWalletBalancesInput = {
-        hin: BigInt(input.hin),
-        assetId: input.assetId ? BigInt(input.assetId) : undefined
-      };
-      const res = await client.getWalletBalances(req);
-      setResult(res);
+      const req = BigInt(input.hin);
+      setResult(await client.getNonce(req));
     } catch (e: any) {
       setError(e.message ?? JSON.stringify(e));
     } finally {
@@ -46,14 +39,11 @@ export default function SectionGetWalletBalances({ client }: { client: HibitClie
 
   return (
     <Section
-      title="GetWalletBalances"
+      title="GetNonce"
       form={
         <div className="flex flex-col gap-2">
           <FormField label="HIN(Hibit chain identity number)" error={errors.hin} required>
-            <input type="number" className="input" {...register('hin')} pattern="[0-9]*" />
-          </FormField>
-          <FormField label="AssetId" error={errors.assetId}>
-            <input type="number" className="input" {...register('assetId')} pattern="[0-9]*" />
+            <input type="number" className="input" {...register('hin')} />
           </FormField>
           <button className="btn" onClick={submit} disabled={loading}>
             {loading ? 'Loading...' : 'Submit'}
