+++ example/components/hibit-client-sections/SectionGetOrderTrades.tsx
@@ -1,19 +1,19 @@
-import { MarketSwapInfo } from '../../src';
-import { HibitClient } from '../../src/hibit-client';
-import Section from './Section';
+import { OrderTradeRecord } from '../../../src';
+import { HibitClient } from '../../../src/hibit-client';
+import Section from '../Section';
 import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { object, string } from 'yup';
-import FormField from './FormField';
+import FormField from '../FormField';
 
 const schema = object({
-  marketId: string()
+  orderId: string().required()
 });
 
-export default function SectionGetMarketsSwapInfo({ client }: { client: HibitClient }) {
+export default function SectionGetOrderTrades({ client }: { client: HibitClient }) {
   const [loading, setLoading] = useState(false);
-  const [result, setResult] = useState<Array<MarketSwapInfo> | null>(null);
+  const [result, setResult] = useState<Array<OrderTradeRecord> | null>(null);
   const [error, setError] = useState<string>('');
 
   const {
@@ -29,8 +29,8 @@ export default function SectionGetMarketsSwapInfo({ client }: { client: HibitCli
     setResult(null);
     setError('');
     try {
-      const req = input.marketId ? BigInt(input.marketId) : undefined;
-      setResult(await client.getMarketsSwapInfo(req));
+      const req = input.orderId;
+      setResult(await client.getOrderTrades(req));
     } catch (e: any) {
       setError(e.message ?? JSON.stringify(e));
     } finally {
@@ -40,11 +40,11 @@ export default function SectionGetMarketsSwapInfo({ client }: { client: HibitCli
 
   return (
     <Section
-      title="GetMarketsSwapInfo"
+      title="GetOrderTrades"
       form={
         <div className="flex flex-col gap-2">
-          <FormField label="MarketId" error={errors.marketId}>
-            <input type="number" className="input" {...register('marketId')} />
+          <FormField label="OrderId" error={errors.orderId} required>
+            <input type="text" className="input" {...register('orderId')} />
           </FormField>
           <button className="btn" onClick={submit} disabled={loading}>
             {loading ? 'Loading...' : 'Submit'}
