+++ package.json
@@ -70,6 +70,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
+    "react-router-dom": "^7.5.1",
     "tailwindcss": "^4.0.8",
     "typescript": "^5.7.3",
     "vite": "^6.1.0",
